Python/Python AI Projects/AiProject1/pyproject.toml
[project]
name = "aiproject1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "nump>=5.5.5.5",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
]
